wip: extract common Viewer component (revealManager wrapper)
for some reason I can't rewrite it more accurate for now - HMR gets broken
